Make dtls work without CoAP
the dtls layer used the CoapMessage internally and it was impossible to
use the dtls layer with anything else than with CoAP. This patch makes
it possible to use the dtls layer with different user provided Message
classes.

There is a generic Message which currently just stores the byte[].
There is also a MessageFactory needed to create such a message out of a
byte[]. All the old code and test cases should work like before, but
now it is possible to e.g. use the dtls layer without any other
protocol.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>